requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
twilio
azure-cognitiveservices-speech
python-multipart
torch
torchaudio
numpy
packaging
langgraph